pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "offject"
version = "2.0.0"
description = "Cross-platform multi-architecture binary patching tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ehab Hussein" }
]
keywords = ["binary", "patching", "assembler", "disassembler", "reverse-engineering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Assemblers",
    "Topic :: Software Development :: Disassemblers",
]
dependencies = [
    "keystone-engine>=0.9.2",
    "capstone>=5.0.0",
    "lief>=0.14.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
offject = "offject.cli:main"

[project.urls]
Homepage = "https://github.com/ehabhussein/offject"
Repository = "https://github.com/ehabhussein/offject"

[tool.hatch.build.targets.wheel]
packages = ["src/offject"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
